requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
openai==1.3.7
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
google-api-python-client==2.108.0
python-dotenv==1.0.0
pydantic==2.5.0
speechrecognition==3.10.0
twilio==8.10.0
websockets==12.0
jinja2==3.1.2
aiofiles==23.2.1
python-dateutil==2.8.2 